Extract shared server startup helpers
